fixed weight init functions

+++ Neuronz-2/src/dezzy/neuronz2/arch/init/XavierInitFunc.java
@@ -15,7 +15,7 @@ final class XavierInitFunc implements WeightInitFunc {
 
 	@Override
 	public <T extends ElementContainer<T>> T initialize(final Random random, final Shape<T> shape, final int numInputs, final int numOutputs, final int numWeights) {
-		final double modifier = Math.sqrt(6 / (numInputs + numOutputs));
+		final double modifier = Math.sqrt(6 / ((double)(numInputs + numOutputs)));
 		final IndexedGenerator generator = i -> ((random.nextDouble() * 2) - 1) * modifier;
 		
 		return shape.generate(generator);
